requirements.txt
fastapi>=0.115
uvicorn>=0.34
langchain>=0.3
langchain-openai>=0.3
langchain-community>=0.3
chromadb>=1.0
pypdf>=5.0
openpyxl>=3.1
unstructured>=0.16
pyyaml>=6.0
pydantic-settings>=2.0
python-dotenv>=1.0

# Phase 3: RAGAS metrics (Response Relevancy)
ragas>=0.2,<0.3
sentence-transformers>=3.0
